Started to work on the new project design: -added a folder called sprites that we will need for small and simple logos for specific texts. -added a fonts folder so that we can called the "eyeglass" font -took out panel3 and panel4 -please ignore the fact not all of it is centered I didn't have time to complete what I wanted to
Juan & Kevin
Please take a look at the text boxes and give them watermarks so that the user knows what to type in.
Not sure if the calendars help feel free to play with it.

Justin
Will this work with your database specifications?

Chris
Feel free to move the css as you wish I will pull tomorrow just let me know if you change anything.