All: My goal is to have something aesthically pleasing for Emily tomorrow. I have decided with Kevin it is best to use bootstrap to accomplish this goal. I added a copy of "about.aspx" and made it "builderproposal_boostrap.aspx" This way we will be able to continue work on either theme.
Kevin, let me know if this bootstrap is outdated or if it is already good to go. Also, if you wouldn't mind setting up the main page I will work on the semantics of design.
Our main goal is the feel of the pinterest "create board" hopefully it has made it to your email. See you later.

-JD